Remove ineffectual if-tests from time integration code for the hydrostatic atmosphere core. After the changes for boundary conditions were added to the trunk, every array has an extra cell, edge, or vertex allocated at index nCells+1, nEdges+1, or nVertices+1, and all occurrences of 0 in indexing array have been replaced with nCells+1, nEdges+1, or nVertices+1.  As a consequence, if-tests such as
   if (cell1 > 0) then

have no effect. Such an if-test could be replaced with

   if (cell1 <= nCells) then

but this is not necessary. We can remove such tests alltogether,
at least in the hydrostatic core; the code in other cores might be
organized such that the bogus cell nCells+1 is used in
computations, in which case some if-tests might still be
necessary.


M    module_time_integration.F


git-svn-id: https://svn-mpas-model.cgd.ucar.edu/trunk/mpas@277 6dc93482-27b8-4c24-8500-90c78a5336cd