Add code in module_io_input.F and module_io_output.F to read and write two new global attributes, on_a_sphere and sphere_radius. If these attributes are not present in an input file, on_a_sphere will default to "YES" and sphere_radius will default to 1.0.
The values of these attributes are accessible in the grid data structure;
that is, they can be accessed in the code as grid % on_a_sphere
(or mesh % on_a_sphere in some places) and grid % sphere_radius.

Note: In the grid data structure, on_a_sphere is a logical, rather than
a character string, as it is stored in the netCDF files.


M    module_grid_types.F
M    module_io_input.F
M    module_io_output.F


git-svn-id: https://svn-mpas-model.cgd.ucar.edu/trunk/mpas@298 6dc93482-27b8-4c24-8500-90c78a5336cd