TRUNK COMMIT
I have added a new subroutine, mpas_init_domain, to mpas_interface.F
in all cores.  It is called from mpas.F

mpas_init_domain, may be used to initialize grid variables that are
not in the netcdf input file.  For example, in the ocean core we
compute maxLevelEdge and maxLevelVertex as derived quantities from a
maxLevelCell variable in the ocean core.  This cannot be done in
mpas_setup_test_case because it needs to be done on restarts.  It
cannot be done in mpas_init because I need to do a halo update on
these variables, and mpas_init is called by block.



git-svn-id: https://svn-mpas-model.cgd.ucar.edu/trunk/mpas@576 6dc93482-27b8-4c24-8500-90c78a5336cd