add option to scale del2 and del4 viscosity based on mesh grid spacing
the option is turned on with "config_h_ScaleWithMesh = .true." in namelist (default is false)

the scaling is done based on meshDensity that is now a part of Registry. This field needs to be in the grid.nc file for the option to work.

the current scaling changes the del4 parameter by a factor of 10 for each doubling of grid spacing.

note, when using this option the :config_h_mom_eddy_visc2" and "config_h_mom_eddy_visc4" parameters should be set to appropriately for the finest region of the mesh. The scaling will increase these parameters in the coarser regions.



git-svn-id: https://svn-mpas-model.cgd.ucar.edu/trunk/mpas@841 6dc93482-27b8-4c24-8500-90c78a5336cd