TRUNK COMMIT
This is to fix a bug in the shallow water code. For the calculations of the visc4 diffusion, vorticity and divergence need to updated on the halos. In the orginal version of the code, the time_level 2 of these variables are updated, which is not necessary, and wrong. This has been corrected, i.e. these two variables of provis are now being updated. 

A test is done on the 10242-cell global quasi-uniform grid, using 8 partitions, and with the config_h_mom_visc4 set to 10000000000000.0 so that the visc4 diffusion is actually turned on. In the result I did not see any anomalies. 

This bug has been confirmed by Mark Petersen, who also pointed out that the correct code should be exactly the same as that in the ocean core. I verified that the codes in questioning are exactly the same in the ocean and sw cores.  



git-svn-id: https://svn-mpas-model.cgd.ucar.edu/trunk/mpas@1256 6dc93482-27b8-4c24-8500-90c78a5336cd