TRUNK COMMIT: Merge branch imp_vert_mix_branch_err with trunk.
This branch corrected the error: the coefficient A in tridiagonal
solve had divisor h(k-1), should be h(k).  This does not produce a
bit-for-bit match, as expected.

Branch was tested by Mark using 120km global grid (p89v-x).

Branch reviewed by Qingshan, with the following comments:

I have finished my review of the imp_vert_mix_error branch. The code looks fine to me, and the results also appear fine (read "not unexpected"). I think for the purpose of fixing the coefficients, this branch is ready to be merged to the trunk. The mass-loss issue is a separate one, and can be addressed in a separate effort. 

For the review, I did the following:
1) The corrected vmix, with rk4 dt = 180, on 125km channel grid.
2) The corrected vmix, with split_explicit dt = 2880, on 125km channel grid;
3) The incorrect vmix, with rk4 dt = 180, on 125km channel grid;
4) The incorrect vmix, with split_expplicit dt = 2880, on 125km channel grid.

I looked at the evolution of energy, enstrophy and mass. Among the test cases, 1) and 3) produce different but similar results. The mass-loss issue is apparent. 2) and 4) produce different but similar results. Mass is conserved.



git-svn-id: https://svn-mpas-model.cgd.ucar.edu/trunk/mpas@1888 6dc93482-27b8-4c24-8500-90c78a5336cd