merge branch to trunk: restart_reproducibility
The following three changes were required to have bit-for-bit restartibility:

1. Call diagnostics before implicit vertical mixing (IVM)

Variables like rho, h_edge, and ke_edge must be updated at the end of
a timestep but before IVM.  Right now we just added an additional call
to ocn_diagnostics_solve just before IVM.  After kpp is added, we
could improve performance by only computing the diagnostics required
for that particular IVM scheme beforehand.  Diagnostics must be
recomputed after IVM as well, since IVM alters u and tracers.

This item prevented bit restartibility for RK4 and split explicit.

2. Compute wTop with the correct advection velocity

There are two velocities for advection:
u, used by momentum equation
uTransport = u + uCorr + uBolus, used by h and tracer equation

wTop in the vertical advection must be computed with the correct
horizontal velocity for each conservation equation.  In this revision,
I changed wTop to accept u as an input variable.  wTop is called with
the correct advection velocity just before computing the tendancies.

This item prevented bit restartibility for split explicit because wTop
was always computed with uTransport, except upon restart when it used
u.

3. Add uBtr to restart file.

In split explicit timestepping, the velocity u is not resplit at the
beginning of each timestep.  Instead, we keep uBtr from the previous
timestep.  Note that uBcl may now include a barotropic component,
because the weights h have changed.  That is OK, because the
GBtrForcing variable subtracts out the barotropic component from the
baroclinic.

This item prevented bit restartibility for split explicit because uBtr
was computed upon restart.  In could be solved by freshly splitting u
at the beginning of each timestep.  However, Stage 1 is set up to take
care of that with the G term, and the additional splitting is not
needed.  Instead, we chose to save uBtr to the restart file for bit
restartability, since this is only the addition of a 2D field.



git-svn-id: https://svn-mpas-model.cgd.ucar.edu/trunk/mpas@2273 6dc93482-27b8-4c24-8500-90c78a5336cd